w7/requirements.txt
connexion[swagger-ui]>=2.6.0,<3.0.0
swagger-ui-bundle>=0.0.2
python_dateutil>=2.6.0
pymongo>=4.6.0
python-dotenv>=1.0.0
Flask>=2.3.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0
six>=1.16.0
